Git prompt python rewrite failed
Starting the Python interpreter takes as long as the bash monstrosity
takes to run. I could rewrite it in a faster language or use Cython, but
one of the big goals in rewriting it was to make it more maintainable,
and none of those options are really simpler than the current solution.