Fix timer workflow launchd prompt preventing the workflow from running
The timer workflow appears to want to prompt me on first install to turn
on a launchd thing to remember my timers when my computer restarts.
There's a problem in this code that makes it happen every time I run the
workflow, and the prompt blocks the rest of the workflow from executing.
Since I don't want to enable this feature, simply remove it.