Fix "faketty: command not found" error in cache cmd
I'm not sure when it happened, but at some point the cache command
stopped working. Something about the way printf was quoting the string
was causing it to not be able to find the faketty alias. I'm not sure
why I used printf in the first place though, since I don't see any risk
of acciental evaluation. The fix then is just to stop using printf.